refactor(fe): split AssignmentAccordion into separate item and container files
